pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "A reflective research agent with tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "arxiv>=2.4.0",
  "httpx>=0.28.1",
  "pydantic-ai>=1.39.0",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "python-dotenv>=1.2.1",
  "rich>=14.2.0",
  "structlog>=25.5.0",
  "tavily-python>=0.7.17",
  "wikipedia-api>=0.8.1",
]

[project.scripts]
research-agent = "src.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
